pyproject.toml
[tool.poetry]
name = "fg-srv"
version = "0.0.1"
description = "fleetG Tasks"
authors = ["Diwaker, diwaker@dj.dj"]
readme = "README.md"
packages = [
    { include = "fg_srv", from = "src/fg_srv" },
    { include = "rest_api", from = "src/fg_srv" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.21.1"
requests = "^2.31.0"
pydantic = "^1.10.8"
paho-mqtt = "^1.6.1"
jinja2 = "^3.0.0"
aiofiles = "^0.6.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
